% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Rice.R
\name{DB_Rice}
\alias{DB_Rice}
\title{Banco de dados.}
\usage{
DB_Rice(Banco_de_dados)
}
\value{
Retorna um banco de dados.
}
\description{
Carrega um Banco de dados que estão armazenada no servidor do SQL.
}
\note{
Vale observar que ao ativar essa função o servidor do banco de dados será
aberta e com isso, não se esqueça de desconectar do servidor.
}
\examples{
DB_Rice('DB_Base') |> head()
}
