% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Treino_Teste.R
\name{DB_Treino_Teste}
\alias{DB_Treino_Teste}
\title{Banco de dados Treino e Teste.}
\format{
Uma lista contendo dois banco de dados.
}
\usage{
DB_Treino_Teste(semente, df, porc_train, porc_test)
}
\arguments{
\item{semente:}{o número da semestre.}

\item{df:}{o banco de dados que devera ser subdividido.}

\item{porc_train:}{a porcentagem de observações do banco de dados que
deve ser realocado no banco de dados treino.}

\item{porc_test:}{a porcentagem de observações do banco de dados que
deve ser realocado no banco de dados teste.}
}
\value{
Retorna uma lista que contém dois banco de dados, treino e teste.
}
\description{
Essa função tem como objetivo de subdividir o banco de dados original em dois,
Banco de dados treino e teste. Alias, vale reparar que deve-se definir a
porcentagem de dados que deve ser colocadas nos dois subconjunto de dados.
}
\examples{
DB_Treino_Teste(2024, DB_Base |> head(50), 0.8, 0.2)

}
