% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_Rice.R
\name{Server_Rice}
\alias{Server_Rice}
\title{função que abre o servidor SQL que hospedam os banco de dados.}
\format{
Servidor SQL
}
\usage{
Server_Rice()
}
\value{
Um servidos SQL
}
\description{
Uma função que abre o servidor SQL que por sua vez, existe diversos banco de dados
sobre o Arroz.
}
\note{
não esqueça de deSconectar do servidor do sql
}
