% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Rice.R
\name{DB_Rice}
\alias{DB_Rice}
\title{Banco de dados.}
\usage{
DB_Rice(Banco_de_dados)
}
\arguments{
\item{Banco_de_dados:}{Banco de dados que queria ler do Servidor SQL do pacote.}
}
\value{
Retorna um banco de dados.
}
\description{
Carrega um Banco de dados que estão armazenada no servidor do SQL do pacote.
}
\note{
Vale observar que ao ativar essa função, o servidor SQL do banco de dados será
aberta.
}
\examples{
DB_Rice('DB_Base') |> head()
}
